Update post route "/" to add tags to new events.
Now takes the tags out of req.body and puts them into variable tags.
After adding new it takes its id and loops through the tags that are in
an array. In each loop it performs AddTag().
The current implementation is not free of bugs. If there are not any
tags in req.body, error with code 400 will be returned but a new event
has still been added. The tags will not be added in this case.